pyproject.toml
[project]
name = "election2024"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Tim Kellogg", email = "timothy.kellogg@gmail.com" }
]
dependencies = [
    "requests>=2.32.3",
    "dspy>=2.5.25",
    "duckduckgo-search>=6.3.3",
    "python-dotenv>=1.0.1",
    "anthropic>=0.39.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/election2024"]
